Added counter and made observer public
