+++ geode-core/src/main/java/org/apache/geode/cache/ProxyClientRequestObserverHolder.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.internal.cache;
+package org.apache.geode.cache;
 
 import java.util.concurrent.atomic.AtomicReference;
