+++ geode-core/src/main/java/org/apache/geode/cache/server/CacheServer.java
@@ -519,4 +519,9 @@ public interface CacheServer {
    */
   Set<InterestRegistrationListener> getInterestRegistrationListeners();
 
+  default void incRejectedProxyRequests() {}
+
+  default long getRejectedProxyRequests() {
+    return 0;
+  }
 }
